+++ src/main/java/kr/hhplus/be/server/domain/concert/repository/ConcertSeatRepository.java
@@ -2,14 +2,13 @@ package domain.concert.repository;
 
 import domain.concert.entity.ConcertSeat;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ConcertSeatRepository {
 
-    List<ConcertSeat> getSeatList(BigInteger concert_date_id);
+    List<ConcertSeat> getSeatList(long concert_date_id);
 
-    ConcertSeat getSeat(BigInteger seat_id);
+    ConcertSeat getSeat(long seat_id);
 
     int updateSeat(ConcertSeat concertSeat);
 }
