+++ src/main/java/kr/hhplus/be/server/domain/concert/repository/UserRepository.java
@@ -2,10 +2,8 @@ package domain.concert.repository;
 
 import domain.concert.entity.Balance;
 
-import java.math.BigInteger;
-
 public interface UserRepository {
-    Balance getBalance(BigInteger user_id);
+    Balance getBalance(long user_id);
 
     int setBalance(Balance balance);
 }
