+++ src/main/java/kr/hhplus/be/server/domain/concert/service/ConcertSeatService.java
@@ -3,16 +3,15 @@ package domain.concert.service;
 import domain.concert.entity.ConcertSeat;
 import domain.concert.entity.User;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ConcertSeatService {
 
-    List<ConcertSeat> getSeatList(BigInteger concert_date_id);
+    List<ConcertSeat> getSeatList(long concert_date_id);
 
-    ConcertSeat updateSeat(User user, BigInteger seat_id);
+    ConcertSeat updateSeat(User user, long seat_id);
 
-    ConcertSeat getConcertSeat(BigInteger seat_id);
+    ConcertSeat getConcertSeat(long seat_id);
 
     void setStatus(ConcertSeat concertSeat);
 }
