+++ src/main/java/kr/hhplus/be/server/domain/concert/service/ConcertService.java
@@ -4,12 +4,11 @@ import domain.concert.entity.Concert;
 import domain.concert.entity.ConcertDate;
 import domain.concert.entity.User;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ConcertService {
 
     List<Concert> getConcert(String name);
 
-    List<ConcertDate> getDate(BigInteger concertId, User user);
+    List<ConcertDate> getDate(long concertId, User user);
 }
