+++ src/main/java/kr/hhplus/be/server/domain/concert/service/UserService.java
@@ -2,11 +2,9 @@ package domain.concert.service;
 
 import domain.concert.entity.Balance;
 
-import java.math.BigInteger;
-
 public interface UserService {
 
-    Balance getBalance(BigInteger user_id);
+    Balance getBalance(long user_id);
 
-    void setBalance(BigInteger user_id, long amount);
+    void setBalance(long user_id, long amount);
 }
